マイグレーション修正

+++ database/migrations/2023_09_03_175220_alter_pets_table.php
@@ -1,5 +1,4 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -16,27 +15,25 @@ return new class extends Migration
             $table->string('age')->nullable()->change();
             $table->string('price')->nullable()->change();
             $table->string('size')->nullable()->change();
-        });
 
-        Schema::table('pets', function (Blueprint $table) {
-            // 新しいカラムの追加
-            $table->string('morph')->nullable()->after('species');
-        });
+            // 新しいカラムの追加前に存在チェック
+            if (!Schema::hasColumn('pets', 'morph')) {
+                $table->string('morph')->nullable()->after('species');
+            }
 
-        Schema::table('pets', function (Blueprint $table) {
-            $table->string('sex')->nullable()->after('species');
-        });
+            if (!Schema::hasColumn('pets', 'sex')) {
+                $table->string('sex')->nullable()->after('species');
+            }
 
-        Schema::table('pets', function (Blueprint $table) {
-            $table->string('weight')->nullable()->after('size');
-        });
+            if (!Schema::hasColumn('pets', 'weight')) {
+                $table->string('weight')->nullable()->after('size');
+            }
 
-        Schema::table('pets', function (Blueprint $table) {
-            $table->date('arrival_day')->nullable()->after('size');
-        });
+            if (!Schema::hasColumn('pets', 'arrival_day')) {
+                $table->date('arrival_day')->nullable()->after('size');
+            }
 
-        // 名前変更
-        Schema::table('pets', function (Blueprint $table) {
+            // 名前変更
             $table->renameColumn('text', 'description');
         });
     }
@@ -51,10 +48,21 @@ return new class extends Migration
             $table->renameColumn('description', 'text');
 
             // 新しいカラムを削除
-            $table->dropColumn('morph');
-            $table->dropColumn('sex');
-            $table->dropColumn('weight');
-            $table->dropColumn('arrival_day');
+            if (Schema::hasColumn('pets', 'morph')) {
+                $table->dropColumn('morph');
+            }
+
+            if (Schema::hasColumn('pets', 'sex')) {
+                $table->dropColumn('sex');
+            }
+
+            if (Schema::hasColumn('pets', 'weight')) {
+                $table->dropColumn('weight');
+            }
+
+            if (Schema::hasColumn('pets', 'arrival_day')) {
+                $table->dropColumn('arrival_day');
+            }
 
             // カラムの変更を元に戻す
             $table->string('size')->nullable(false)->change();
@@ -63,4 +71,3 @@ return new class extends Migration
         });
     }
 };
-
